% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mockaRoo-package.r
\docType{package}
\name{mockaRoo}
\alias{mockaRoo}
\alias{mockaRoo-package}
\title{Generate mock data with the Mockaroo API}
\description{
Allows the generation of mock or test data
via API calls to Mockaroo.com
}

